feat : interceptor에서 request header X-USER-KEY 검사하도록 추가

+++ src/main/java/org/c4marathon/assignment/controller/FileController.java
@@ -2,14 +2,23 @@ package org.c4marathon.assignment.controller;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.c4marathon.assignment.dto.FileDownloadResponseDto;
+import org.c4marathon.assignment.dto.ResponseDto;
+import org.c4marathon.assignment.exception.FileEmptyException;
+import org.c4marathon.assignment.exception.IllegalParameterException;
 import org.c4marathon.assignment.service.FileDeleteService;
+import org.c4marathon.assignment.service.FileDownloadService;
 import org.c4marathon.assignment.service.FileUploadService;
+import org.c4marathon.assignment.exception.ParameterEmptyException;
+import org.springframework.http.HttpHeaders;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
 
 import java.io.IOException;
+import java.io.UnsupportedEncodingException;
+import java.net.MalformedURLException;
 
 @Controller
 @RequiredArgsConstructor
@@ -17,19 +26,38 @@ import java.io.IOException;
 public class FileController {
     private final FileUploadService fileUploadService;
     private final FileDeleteService fileDeleteService;
-
+    private final FileDownloadService fileDownloadService;
     @PostMapping("/file")
-    public ResponseEntity<Object> uploadFile(@ModelAttribute("file") MultipartFile file,
-                                             @RequestHeader("user_key") String userKey) throws IOException {
-        fileUploadService.uploadFileHandler(file, userKey);
-        return ResponseEntity.ok().body(file.getOriginalFilename() + " upload success");
+    public ResponseEntity<ResponseDto> uploadFile(@RequestParam(value = "file", required = false) MultipartFile[] fileList,
+                                                  @RequestHeader(value = "X-USER-KEY") String userKey) throws IOException {
+        if(fileList == null || fileList.length == 0){
+            throw new FileEmptyException();
+        }
+        ResponseDto responseDto = fileUploadService.uploadFileListHandler(fileList, userKey);
+        return ResponseEntity.ok().body(responseDto);
     }
 
     @DeleteMapping("/file")
-    public ResponseEntity<Object> deleteFile(@RequestParam("file-id") Long fileId,
-                                             @RequestHeader("user_key") String userKey) throws IOException {
-        fileDeleteService.deleteFileHandler(userKey, fileId);
-        return ResponseEntity.ok().body("delete success");
+    public ResponseEntity<ResponseDto> deleteFile(@RequestParam(value = "file-id", required = false) Long fileId,
+                                             @RequestHeader("X-USER-KEY") String userKey) throws IOException {
+        if(fileId == null){
+            throw new ParameterEmptyException();
+        }
+        ResponseDto responseDto = fileDeleteService.deleteFileHandler(userKey, fileId);
+        return ResponseEntity.ok().body(responseDto);
+    }
+
+    @GetMapping("/file")
+    public ResponseEntity<Object> downloadFile(@RequestParam(value = "file-id", required = false) Long fileId,
+                                               @RequestHeader("X-USER-KEY") String userKey) throws MalformedURLException, UnsupportedEncodingException {
+        if(fileId == null){
+            throw new ParameterEmptyException();
+        }
+        FileDownloadResponseDto fileDownloadResponseDto = fileDownloadService.fileDownloadHandler(userKey, fileId);
+        return ResponseEntity.ok()
+                .header(HttpHeaders.CONTENT_DISPOSITION, fileDownloadResponseDto.getContentDisposition())
+                .header(HttpHeaders.CONTENT_TYPE, fileDownloadResponseDto.getContentType())
+                .body(fileDownloadResponseDto.getResource());
     }
 }
